build.rs fix
